Search box position / Focus testing
Updated search box to have an absolute position so it remains at the top of the page if user scrolls down. Makes it fee like its part of the banner - we can go the other route also and make it fixed  and detach from the banner completely. Even overlaying the full screen if needed. Added a blur cuz why not.

Also, tried to follow svelte design patterns and moved adding event listeners to the <svelte:window ..> event ( <svelte:document ...> didn't work ). I cleared out an old function as listener will handle. Added a different function to clear and close the search that is called when clicking a searched link.