Search Feature (#51)
* Init Search branch

Attempt to add search feature based on https://joyofcode.xyz/blazing-fast-sveltekit-search

* tweak css

* tweak layout

* Fixed logo being hidden in App.css

* Revert "tweak layout"

This reverts commit 2e5b3976a157d9400d4fe3c5ca070974d3de3342.

* Changed to Popover

Now using Popover API to show search window, added onClick to links in search to togglepopover and to clear the search term (though we maybe we keeping search term in there is cool?)

* tweak css

* fix imports, some formatting and null checks

* focus input on popover

* Search box position / Focus testing

Updated search box to have an absolute position so it remains at the top of the page if user scrolls down. Makes it feel like it's part of the banner - we can go the other route also and make it fixed and detach from the banner completely. Even overlaying the full screen if needed. Added a blur cuz why not.

Also, tried to follow svelte design patterns and moved adding event listeners to the <svelte:window ..> event ( <svelte:document ...> didn't work ). I cleared out an old function as listener will handle. Added a different function to clear and close the search that is called when clicking a searched link.

---------

Co-authored-by: PAUL HEBERT <pdhebert@us.ibm.com>
Co-authored-by: teauxfu <alex@southsun.tech>